Add a basic lexer implementation
This implementation is inspired by Rob Pike's talk on creating a
concurrent lexer.

This version of the implementation already supports some easy to
implement tokens, the literal text that gives rise to the token and its
location in the source.

Among the tokens implemented are the comment token, the name (attribute)
token and the end-of-file token.